Add files via upload
OpenSCAD code for most recent version of the 3D printable Ethology robots.  Designed to be printable on a LulzBot mini 6" x 6" build plate.  The Grey and Orange folders contain STLs of parts currently printed in those colors.  The Ninja Flex folder contains STL for the sleeve that may go over a bump sensor.  However, weather stripping is currently recommended to be glued over the sensor such that the bubble or tube part of the stripping is over the resistor part of the sensor.  The castor file is not in a color folder, but will print the castors that go on either end of the robot.

For structural integrity, it is recommended that the "combined" version of the bumper STL be printed.  This combines the extension and bumper proper into one part with one less point of failure per bumper.  The tire may be printed in Ninja Flex if traditional Parallax tires do not seem to provide enough traction in order to activate the bump sensors.  Otherwise, this may be omitted from the robot design.

For questions, please contact Nick Livingston using the gm@!l account of the same name as this GitHub username.